修正了纹理坐标颠倒的问题

+++ Path-Tracing/vector.h
@@ -59,4 +59,4 @@ public:
 	}
 };
 
-typedef Vec3<float> Vec3f;
+typedef Vec3<double> Vec3f;
